examples/streamlit/requirements.txt
streamlit>=1.47.1
langgraph>=0.5.4
langchain-mcp-adapters>=0.1.9
langchain>=0.3.27
langchain-openai>=0.3.28
python-dotenv>=1.0.0
